% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armorPen.R
\name{armorPen}
\alias{armorPen}
\title{Total Armor Penetration Calculator}
\usage{
armorPen(flatRed, pRed, pPen, leth)
}
\arguments{
\item{flatRed}{Flat armor reduction. May be taken from already-known total.}

\item{pRed}{Total Percent armor reduction. May be taken from already-known total or calculated using pRedCalc.}

\item{pPen}{Total Percent armor penetration. May be taken from already-known total or calculated using pRedCalc.}

\item{leth}{Total Percent armor reduction. May be taken from already-known total or calculated using pRedCalc.}
}
\value{
A vector containing three values.
First value returns a flat reduction amount which for damage calculations must be subtracted from theoretical armor amount.
Second value returns total percent armor Reduction input for purposes of organization.
Third value repeats second value purpose but for total percent Penetration.
}
\description{
Combines calculated values of flat armor reduction, % armor reduction, % armor penetration, and lethality into a single vector.It also checks to make sure inputted variables are valid.
}
\examples{
armorPen(25,.12,.20,15)
}
