% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSpeed.R
\name{aSpeed}
\alias{aSpeed}
\title{Attack Speed Calculator}
\usage{
aSpeed(ratio, modifier, ...)
}
\arguments{
\item{ratio}{Base attack speed ratio of champion, is always equal to base attack speed.}

\item{modifier}{Attack speed modifier.}

\item{...}{any further modifiers derived from skills or items.}
}
\value{
A value indicating total attack speed (attacks per second).
}
\description{
Function designed to calculate attack speed of a champion given certain bonuses.
}
\examples{
aSpeed(.75,.25,.3,.25)
}
