% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdReduction.R
\name{cdReduction}
\alias{cdReduction}
\title{Cool down Reduction}
\usage{
cdReduction(haste1, ...)
}
\arguments{
\item{haste1}{first haste stat provided by item, mastery, champion ability, etc.}

\item{...}{Any further additional haste bonuses.}
}
\value{
Total percent cool down reduction.
}
\description{
Calculates total cool down reduction of abilities given by the haste item stat.
}
\examples{
cdReduction(20,12,15)
}
