% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lethCalc.R
\name{lethCalc}
\alias{lethCalc}
\title{Lethality Calculator}
\usage{
lethCalc(lethality, level)
}
\arguments{
\item{lethality}{base lethality given from item}

\item{level}{current level of champion being analyzed in scenario}
}
\value{
Total lethality, to be used as part of armorPen() calculation.
}
\description{
Function designed to calculate applied lethality of a champion based on total lethality gained according to the current level of a champion. As champion level increases, the ratio at which lethality translates to flat armor penetration also increases.
}
\examples{
lethCalc(30,16)
}
