% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolscrape.R
\name{lolscrape}
\alias{lolscrape}
\title{Champion Data Scraper}
\usage{
lolscrape()
}
\value{
A data table containing current champion base statistics. Must be assigned.
}
\description{
Creates data table by scraping current champion base statistics from the wiki.
}
\examples{
champData <- lolscrape()
}
