% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpeed.R
\name{mSpeed}
\alias{mSpeed}
\title{Movement Speed Calculator}
\usage{
mSpeed(base, fmodifier, pmodifier, ...)
}
\arguments{
\item{base}{base attack speed ratio of champion, is always equal to base attack speed.}

\item{fmodifier}{Flat movement speed modifier.}

\item{pmodifier}{Percentage-based movement speed modifier.}

\item{...}{Any further percentage-based modifiers derived from skills or items.}
}
\value{
A value indicating total movement speed of a given champion with current modifiers.
}
\description{
Function designed to calculate attack speed of a champion given certain bonuses.
}
\examples{
mSpeed(325, 45, .07, .07)
}
