% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRedCalc.R
\name{pRedCalc}
\alias{pRedCalc}
\title{Percent Reduction (Armor)}
\usage{
pRedCalc(champAbility, BC)
}
\arguments{
\item{champAbility}{Percent armor reduction based on a specific champion ability.}

\item{BC}{Number of stacks of black cleaver procs being used in calculation. Limit 6..}
}
\value{
Total percentage of armor reduced for purposes of damage calculations.
}
\description{
Calculates percent armor reduction based on two factors: champion's ability and Black Cleaver stacks.
}
\examples{
pRedCalc(.25,6)
}
